Constant line detail independent from zoom
Create a reference point attribute for all materials that defines a
reference point that is used for the distance calcuation to determine
the line-zoom factor.

Since we can now control those values per triangular face, but also per
quad we reduce the artifacts of earlier attempts directly using the
world position.